debt: playground: extract live-history + display-update from main.js
Two tightly-coupled modules landed together (circular imports are
fine for runtime-evaluated functions).

- main/live-history.js: fetchLiveHistory, loadLiveHistoryIntoStore,
  recordLiveHistoryPoint + window.__getHistoryPointCount test hook.
  Calls back into display-update.rerenderWithHistoryFallback after
  the /api/history fetch resolves.
- main/display-update.js: updateDisplay + trend/formatting helpers +
  rerenderWithHistoryFallback. Owns schematicHandle, lastState,
  lastResult, liveFrameSeen, yesterdayHigh, confirmedYesterdayHigh,
  lastDay. Exposes setSchematicHandle/getLastFrame/setLiveFrameSeen/
  resetYesterdayTracking for main.js to wire through.

main.js: exports `running` (read by display), removes all the
display-owned state declarations, replaces direct writes with
the accessor calls.

main.js: 1627 → 1091 lines. npm test:unit 782/782 pass. knip clean.
